refactor: 통합테스트 패키지를 controller -> integration으로 수정
