+++ pennyway-app-external-api/src/test/java/kr/co/pennyway/api/apis/auth/integration/OAuthControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package kr.co.pennyway.api.apis.auth.controller;
+package kr.co.pennyway.api.apis.auth.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.co.pennyway.api.apis.auth.dto.PhoneVerificationDto;
