+++ pennyway-app-external-api/src/test/java/kr/co/pennyway/api/apis/auth/integration/UserAuthControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package kr.co.pennyway.api.apis.auth.controller;
+package kr.co.pennyway.api.apis.auth.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.Cookie;
